+++ lib/Material/Routes/SubareasMaterial.dart
@@ -1,29 +1,41 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:rock_carrot/Baseitems/BaseItems.dart';
-import 'package:rock_carrot/Baseitems/Countries.dart';
+import 'package:rock_carrot/Baseitems/Areas.dart';
+import 'package:rock_carrot/Baseitems/Subareas.dart';
 import 'package:rock_carrot/Baseitems/cubit/update_cubit.dart';
-import 'package:rock_carrot/Material/BaseItemTile.dart';
-import 'package:rock_carrot/Material/BaseMaterial.dart';
+import 'package:rock_carrot/Material/Tiles/BaseItemTile.dart';
+import 'package:rock_carrot/Material/Routes/BaseMaterial.dart';
+import 'package:rock_carrot/Material/ProgressNotifier.dart';
 import 'package:rock_carrot/Material/Snackbar.dart';
 import 'package:rock_carrot/Web/Sandstein.dart';
 import 'package:rock_carrot/Web/SandsteinSql.dart';
+import 'package:rock_carrot/Web/Teufelsturm.dart';
 
-class CountryMaterial extends StatefulWidget {
+class SubAreasMaterial extends StatefulWidget {
+  final Area _parentItem;
+  // support updateing the child Values
+  final ProgressNotifier _parentProgressNotifier;
+  final UpdateCubit _parentUpdateCubit;
+
+  SubAreasMaterial(
+      this._parentItem, this._parentProgressNotifier, this._parentUpdateCubit);
+
+  // transfer country to state object
   @override
-  _CountryMaterialState createState() => _CountryMaterialState();
+  _SubAreasMaterialState createState() {
+    return _SubAreasMaterialState(_parentItem);
+  }
 }
 
-class _CountryMaterialState
-    extends BaseItemsMaterialStatefulState<CountryMaterial> {
-  /// All basic functionality is in this object
-  final Countries _countries = Countries();
+class _SubAreasMaterialState
+    extends BaseItemsMaterialStatefulState<SubAreasMaterial> {
+  /// All basic functionality is in this object (incl. parentItem)
+  final Subareas _subareas;
 
-  _CountryMaterialState() {
-    searchBar = initializeSearchBar(BaseItem(
-      name: 'Countries',
-      childCountInt: -1,
-    ));
+  _SubAreasMaterialState(Area area) : _subareas = Subareas(area) {
+    searchBar = initializeSearchBar(_subareas.parentArea);
+    // default sorting ist by child count
+    sortAlpha = false;
   }
 
   /// build the Scaffold
@@ -34,23 +46,29 @@ class _CountryMaterialState
         // enable Refresh data with pulldown
         body: RefreshIndicator(
           onRefresh: () async {
+            int count;
             try {
-              await Sandstein().updateCountries();
+              count = await Sandstein()
+                  .updateSubareasInclComments(_subareas.parentArea);
             } catch (e) {
               ScaffoldMessenger.of(context)
                   .showSnackBar(ErrorSnack(e.toString()));
+              count = 0;
             }
+
+            // update parent tile
+            widget._parentProgressNotifier.setStaticValue(count);
+            widget._parentUpdateCubit.callGetValueAsync(_subareas.parentArea);
 
             setState(() {});
           },
-          child: FutureBuilder<List<Country>>(
+          child: FutureBuilder<List<Subarea>>(
             builder: futureBuildItemList,
-            future: _countries.getCountries(),
+            future: _subareas.getSubareas(),
           ),
         ));
   }
 
-  /// generate Listview with all Country Data
   @override
   Widget buildItemList(AsyncSnapshot snapshot) {
     // store snapshot data in local list
@@ -65,12 +83,12 @@ class _CountryMaterialState
         },
       );
     }
-    // build the Widget
+
     return ListView.builder(
       padding: EdgeInsets.all(0),
       itemCount: baseitem_list.length,
       itemBuilder: (context, i) {
-        final country = baseitem_list[i] as Country;
+        final subarea = baseitem_list[i] as Subarea;
         return Column(children: [
           // only first time generate a devider
           (i == 0)
@@ -79,13 +97,13 @@ class _CountryMaterialState
                   thickness: 1,
                 )
               : Container(),
-          // Cubit for updating Timestamp
           BlocProvider(
             create: (context) => UpdateCubit(),
             child: BaseItemTile(
-              country,
-              updateFunction: Sandstein().updateAreas,
-              deleteFunction: Sandstein().deleteAreasFromDatabase,
+              subarea,
+              updateFunction: Sandstein().updateRocksIncludingSubitems,
+              updateAllFunction: Teufelsturm().updateTTComments,
+              deleteFunction: Sandstein().deleteRocksFromDatabase,
             ),
           ),
           Divider(
